+++ app/code/local/Radarsofthouse/Reepay/Block/Info/Swish.php
@@ -9,14 +9,14 @@
  * @license     http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  *
  */
-class Radarsofthouse_Reepay_Block_Info_Klarna extends Mage_Payment_Block_Info
+class Radarsofthouse_Reepay_Block_Info_Swish extends Mage_Payment_Block_Info
 {
     protected $_instructions;
 
     protected function _construct()
     {
         parent::_construct();
-        $this->setTemplate('reepay/info/klarna.phtml');
+        $this->setTemplate('reepay/info/swish.phtml');
     }
 
     public function getInstructions()