+++ app/code/local/Radarsofthouse/Reepay/controllers/WebhooksController.php
@@ -160,6 +160,9 @@ class Radarsofthouse_Reepay_WebhooksController extends Mage_Core_Controller_Fron
                 $transactionID = Mage::helper('reepay')->addCaptureTransactionToOrder($order, $reepayTransactionData);
 
                 $_autoCapture = Mage::helper('reepay')->getConfig('auto_capture', $order->getStoreId());
+                if( $order->getPayment()->getMethodInstance()->getCode() == "reepay_swish" ){
+                    $_autoCapture = true;
+                }
                 if($_autoCapture){
 
                     $charge = Mage::helper('reepay/charge')->get($apiKey, $orderId);